Make ProfileFragment.
