+++ app/src/main/java/me/zhanghai/android/douya/profile/ui/ProfileActivity.java
@@ -8,6 +8,7 @@ package me.zhanghai.android.douya.profile.ui;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -35,35 +36,20 @@ import me.zhanghai.android.douya.network.api.info.frodo.Review;
 import me.zhanghai.android.douya.network.api.info.frodo.UserItems;
 import me.zhanghai.android.douya.profile.content.ProfileResource;
 import me.zhanghai.android.douya.ui.ContentStateLayout;
+import me.zhanghai.android.douya.util.FragmentUtils;
 import me.zhanghai.android.douya.util.LogUtils;
 import me.zhanghai.android.douya.util.ToastUtils;
 import me.zhanghai.android.douya.util.ViewUtils;
 
-public class ProfileActivity extends AppCompatActivity implements ProfileResource.Listener,
-        ProfileHeaderLayout.Listener, ConfirmUnfollowUserDialogFragment.Listener {
+public class ProfileActivity extends AppCompatActivity {
 
-    private static final String KEY_PREFIX = ProfileActivity.class.getName() + '.';
+    private static final String KEY_PREFIX = ProfileFragment.class.getName() + '.';
 
     private static final String EXTRA_USER_ID_OR_UID = KEY_PREFIX + "user_id_or_uid";
     private static final String EXTRA_USER = KEY_PREFIX + "user";
     private static final String EXTRA_USER_INFO = KEY_PREFIX + "user_info";
 
-    @BindView(R.id.scroll)
-    ProfileLayout mScrollLayout;
-    @BindView(R.id.header)
-    ProfileHeaderLayout mHeaderLayout;
-    @BindView(R.id.dismiss)
-    View mDismissView;
-    @BindView(R.id.toolbar)
-    Toolbar mToolbar;
-    @BindView(R.id.contentState)
-    ContentStateLayout mContentStateLayout;
-    @BindView(R.id.content)
-    RecyclerView mContentList;
-
-    private ProfileResource mProfileResource;
-
-    private ProfileAdapter mProfileAdapter;
+    private ProfileFragment mProfileFragment;
 
     public static Intent makeIntent(String userIdOrUid, Context context) {
         return new Intent(context, ProfileActivity.class)
@@ -81,81 +67,29 @@ public class ProfileActivity extends AppCompatActivity implements ProfileResourc
     }
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
 
         overridePendingTransition(0, 0);
 
-        super.onCreate(savedInstanceState);
+        // Calls ensureSubDecor().
+        findViewById(android.R.id.content);
 
-        setContentView(R.layout.profile_activity);
-        ButterKnife.bind(this);
-
-        Intent intent = getIntent();
-        String userIdOrUid = intent.getStringExtra(EXTRA_USER_ID_OR_UID);
-        User user = intent.getParcelableExtra(EXTRA_USER);
-        UserInfo userInfo = intent.getParcelableExtra(EXTRA_USER_INFO);
-
-        CustomTabsHelperFragment.attachTo(this);
-        mProfileResource = ProfileResource.attachTo(userIdOrUid, user, userInfo, this);
-
-        mScrollLayout.setListener(new ProfileLayout.Listener() {
-            @Override
-            public void onEnterAnimationEnd() {}
-            @Override
-            public void onExitAnimationEnd() {
-                finish();
-            }
-        });
-        mScrollLayout.enter();
-
-        mDismissView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                exit();
-            }
-        });
-
-        setSupportActionBar(mToolbar);
-        getSupportActionBar().setTitle(null);
-
-        if (mProfileResource.hasUserInfo()) {
-            mHeaderLayout.bindUserInfo(mProfileResource.getUserInfo());
-        } else if (mProfileResource.hasUser()) {
-            mHeaderLayout.bindUser(mProfileResource.getUser());
-        }
-        mHeaderLayout.setListener(this);
-
-        if (ViewUtils.hasSw600dp(this)) {
-            int columnCount = ViewUtils.isInLandscape(this) ? 3 : 2;
-            mContentList.setLayoutManager(new StaggeredGridLayoutManager(columnCount,
-                    StaggeredGridLayoutManager.VERTICAL));
+        if (savedInstanceState == null) {
+            Intent intent = getIntent();
+            String userIdOrUid = intent.getStringExtra(EXTRA_USER_ID_OR_UID);
+            User user = intent.getParcelableExtra(EXTRA_USER);
+            UserInfo userInfo = intent.getParcelableExtra(EXTRA_USER_INFO);
+            mProfileFragment = ProfileFragment.newInstance(userIdOrUid, user, userInfo);
+            FragmentUtils.add(mProfileFragment, this, android.R.id.content);
         } else {
-            mContentList.setLayoutManager(new LinearLayoutManager(this));
+            mProfileFragment = FragmentUtils.findById(this, android.R.id.content);
         }
-        mProfileAdapter = new ProfileAdapter();
-        mContentList.setAdapter(mProfileAdapter);
-        if (mProfileResource.isLoaded()) {
-            mProfileResource.notifyChangedIfLoaded();
-        } else {
-            mContentStateLayout.setLoading();
-        }
-    }
-
-    // When moved into fragment, this will be needed.
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-
-        mProfileResource.detach();
     }
 
     @Override
     public void onBackPressed() {
-        exit();
-    }
-
-    private void exit() {
-        mScrollLayout.exit();
+        mProfileFragment.onBackPressed();
     }
 
     @Override
@@ -163,88 +97,5 @@ public class ProfileActivity extends AppCompatActivity implements ProfileResourc
         super.finish();
 
         overridePendingTransition(0, 0);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        super.onCreateOptionsMenu(menu);
-
-        getMenuInflater().inflate(R.menu.profile, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onPrepareOptionsMenu(Menu menu) {
-        super.onPrepareOptionsMenu(menu);
-
-        // TODO: Block or unblock.
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_send_doumail:
-                // TODO
-                return true;
-            case R.id.action_blacklist:
-                // TODO
-                return true;
-            case R.id.action_report_abuse:
-                // TODO
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-
-    @Override
-    public void onLoadError(int requestCode, VolleyError error) {
-        LogUtils.e(error.toString());
-        mContentStateLayout.setError();
-        ToastUtils.show(ApiError.getErrorString(error, this), this);
-    }
-
-    @Override
-    public void onUserInfoChanged(int requestCode, UserInfo newUserInfo) {
-        mHeaderLayout.bindUserInfo(newUserInfo);
-    }
-
-    @Override
-    public void onUserInfoWriteStarted(int requestCode) {
-        mHeaderLayout.bindUserInfo(mProfileResource.getUserInfo());
-    }
-
-    @Override
-    public void onUserInfoWriteFinished(int requestCode) {
-        mHeaderLayout.bindUserInfo(mProfileResource.getUserInfo());
-    }
-
-    @Override
-    public void onChanged(int requestCode, UserInfo newUserInfo, List<Broadcast> newBroadcastList,
-                          List<User> newFollowingList, List<Diary> newDiaryList,
-                          List<UserItems> newUserItemList, List<Review> newReviewList) {
-        mProfileAdapter.setData(new ProfileAdapter.Data(newUserInfo, newBroadcastList,
-                newFollowingList, newDiaryList, newUserItemList, newReviewList));
-        mContentStateLayout.setLoaded(true);
-    }
-
-    @Override
-    public void onEditProfile(UserInfo userInfo) {
-
-    }
-
-    @Override
-    public void onFollowUser(UserInfo userInfo, boolean follow) {
-        if (follow) {
-            FollowUserManager.getInstance().write(userInfo, true, this);
-        } else {
-            ConfirmUnfollowUserDialogFragment.show(this);
-        }
-    }
-
-    @Override
-    public void unfollowUser() {
-        FollowUserManager.getInstance().write(mProfileResource.getUserInfo(), false, this);
     }
 }
